# schema/openapi.yaml
...
components:
  schemas:
...
    Annotations:
      properties:
        cryostat:
          additionalProperties:
            type: string
          type: object
        platform:
          additionalProperties:
            type: string
          type: object
      type: object
...
    HeapDump:
      properties:
        downloadUrl:
          type: string
        heapDumpId:
          type: string
        jvmId:
          type: string
        lastModified:
          format: int64
          type: integer
        metadata:
          $ref: '#/components/schemas/Metadata'
        size:
          format: int64
          type: integer
      type: object
    HttpServerResponse:
      properties:
        chunked:
          type: boolean
        statusCode:
          format: int32
          type: integer
        statusMessage:
          type: string
      type: object
...
    JsonObject:
      items:
        properties:
          key:
            type: string
          value: {}
        type: object
      type: array
...
    Metadata:
      properties:
        labels:
          additionalProperties:
            type: string
          type: object
      type: object
...
    Target:
      properties:
        agent:
          readOnly: true
          type: boolean
        alias:
          pattern: \S
          type: string
        annotations:
          $ref: '#/components/schemas/Annotations'
        connectUrl:
          format: uri
          type: string
        id:
          $ref: '#/components/schemas/UUID'
        jvmId:
          type: string
        labels:
          additionalProperties:
            type: string
          type: object
      required:
        - id
        - connectUrl
        - alias
        - labels
        - annotations
      type: object
...
    UUID:
      format: uuid
      pattern: '[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}'
      type: string
...
info:
  contact:
    email: cryostat-development@googlegroups.com
    name: Cryostat Community
    url: https://cryostat.io
  description: Cloud-Native JDK Flight Recorder
  license:
    name: Apache 2.0
    url: https://github.com/cryostatio/cryostat/blob/main/LICENSE
  title: Cryostat API
  version: 5.0.0-snapshot
openapi: 3.1.0
paths:
...
  /api/v5/audit/targets/{jvmId}:
    get:
      parameters:
        - in: path
          name: jvmId
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Target'
          description: OK
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
        - SecurityScheme: []
      summary: Target By Jvm Id
      tags:
        - Audit
...
  /api/v5/targets/{id}:
    get:
      description: |
        Get details about a particular target given its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Target'
          description: OK
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
        - SecurityScheme: []
      summary: Get a target by ID
      tags:
        - Targets
  /api/v5/targets/{jvmId}:
    delete:
      description: |
        Delete a custom target definition by its JVM ID. Only allows deletion of targets that were defined by the
        Custom Target API. Other targets must be removed by the discovery mechanisms which discovered them.
      parameters:
        - in: path
          name: jvmId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
        - SecurityScheme: []
      summary: Delete a custom target by JVM ID
      tags:
        - Custom Discovery
...
  /api/v5/targets/{jvmId}/diagnostics/heap-dump:
    get:
      parameters:
        - in: path
          name: jvmId
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/HeapDump'
                type: array
          description: OK
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
        - SecurityScheme: []
      summary: Get Heap Dumps
      tags:
        - Target Heap Dump Endpoint
    post:
      description: |
        Request the remote target to perform a heap dump.
      parameters:
        - in: path
          name: jvmId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HttpServerResponse'
        required: true
      responses:
        "200":
          content:
            text/plain:
              schema:
                type: string
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
        - SecurityScheme: []
      summary: Initiates a heap dump on the specified target
      tags:
        - Target Heap Dump Endpoint
  /api/v5/targets/{jvmId}/diagnostics/heap-dump/upload:
    post:
      parameters:
        - in: path
          name: jvmId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                heapDump:
                  format: binary
                  type: string
                jobId:
                  type: string
                labels:
                  $ref: '#/components/schemas/JsonObject'
              type: object
        required: true
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
        - SecurityScheme: []
      summary: Upload Heap Dump
      tags:
        - Target Heap Dump Endpoint
  /api/v5/targets/{jvmId}/diagnostics/heap-dump/{heapDumpId}:
    delete:
      parameters:
        - in: path
          name: heapDumpId
          required: true
          schema:
            type: string
        - in: path
          name: jvmId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
        - SecurityScheme: []
      summary: Delete Heap Dump
      tags:
        - Target Heap Dump Endpoint
  /api/v5/targets/{jvmId}/diagnostics/heap-dump/{heapDumpId}/analyze:
    post:
      parameters:
        - in: path
          name: heapDumpId
          required: true
          schema:
            type: string
        - in: path
          name: jvmId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HttpServerResponse'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
        - SecurityScheme: []
      summary: Analyze Heap Dump
      tags:
        - Target Heap Dump Endpoint
...